fix: prevent PID wind-up when VBUS is absent at startup
* fix: prevent PID wind-up when VBUS is absent at startup

SpeedController() now checks vbusVref against VBUS_MIN_THRESHOLD before
running the speed regulator. If VBUS is below the threshold the PID
integrator is reset and speedOutput is held at zero, preventing integrator
wind-up and unintended drive output when the motor power rail is not yet
present.

PID_OUTPUT_MAX and VBUS_MIN_THRESHOLD are exposed as user-settable defines
in config.h. PID_OUTPUT_MAX replaces the previously hardcoded ceiling of
255 in the closed-loop path. VBUS_MIN_THRESHOLD defaults to 96 (~8 V with
the default NEVB-MTR1-C-1 resistor divider).

* chore: update version

* chore: update version

* chore: update PID_OUTPUT_MAX default to 200

* fix: apply scpi.cpp cast

+++ main/scpi.cpp
@@ -413,7 +413,7 @@ static void ConfigureMotorSpeed(SCPI_C commands, SCPI_P parameters, Stream &inte
         scpiParser.last_error = ErrorCode::MissingOrInvalidParameter;
         return;
     }
-    speedInput = ((param * SPEED_CONTROLLER_MAX_INPUT * MOTOR_POLES) >> 3) / ((uint32_t)SPEED_CONTROLLER_MAX_SPEED * 15);
+    speedInput = ((uint32_t)(param * SPEED_CONTROLLER_MAX_INPUT * MOTOR_POLES) >> 3) / ((uint32_t)SPEED_CONTROLLER_MAX_SPEED * 15);
     scpiParser.last_error = ErrorCode::NoError;
 }
 #endif
